worldwidetelescope.org: categorize some URLs as "frontend"
This means that they're part of the "chrome" that builds up the user
experience of a visitor to the general site. Webclient resource files not
included in this category, for the time being, nor are supporting data,
although these categorizations all have some flex. And also I'm not yet
exactly sure what I'm going to do with this categorization, anyway.