worldwidetelescope.org: add miscellaneous static data files
These live in the same `/data/` directory as the files accessible via
`/wwtweb/catalog.aspx`, except that they have nonstandard suffixes so that the
`catalog.aspx` script cannot be used to access to them.

Note that we categorize `catalog.aspx` URLs as "backend" while these are
"frontend". I think the correct solution is that the particular `catalog.aspx`
path should actually be considered a frontend URL, since it just provides
access to static data files.

There were two files in my data backup directory that gave me 404s when I
tried to fetch them; I think this indicates that the files only live on one of
the web servers, and that my requests happened to be going to the other one. They are:

- iss.wwtl
- lang_en.tdf