worldwidetelescope.org: give up on the sitemap for now
It wouldn't be too hard to wire things up to serve a static sitemap
file, but it seems that sitemaps need to contain absolute URLs, and the
protocol matters. So we ought to serve different sitemaps over HTTP and
HTTPS, and that's just too much of a hassle for me to deal with right
now. Especially because the guidance from Google suggests that we don't
really need a sitemap for this domain. So let's punt.